Added days 3 and 4
Finished half of day 3, this one was really tough but I found a formula via google that yielded a solution. Might come back to this one later but I'm not very good at maths.
Day 4 was MUCH more managable, did a very brute-force method. But if it works it works, not going for amazing code on this one :)